requirements.txt
gradio
pillow
pyyaml
tqdm
reportlab
pytest
llama-cpp-python
# optional:
# torch
# transformers
# accelerate
# bitsandbytes
